fix: resolve all clippy warnings for clean CI
- Fix field_reassign_with_default in config tests
- Replace io::Error::new with io::Error::other
- Remove unnecessary clone on Copy type
- Replace vec\! with array literals where appropriate
- Fix excessive precision in float literals
- Replace manual range contains with idiomatic syntax
- Use is_empty() instead of len() >= 1

All clippy warnings resolved with -D warnings flag.

+++ sakurs-core/tests/strategy_comparison_tests.rs
@@ -149,7 +149,7 @@ fn test_memory_constrained_processing() {
     let expected_boundaries = 10100;
     let tolerance = 5;
     let actual_boundaries = result.boundaries.len() as i32;
-    let difference = (actual_boundaries - expected_boundaries as i32).abs();
+    let difference = (actual_boundaries - expected_boundaries).abs();
 
     assert!(
         difference <= tolerance,
